{"schema":"decodex/commit/1","summary":"Split MCP project context resources","authority":"XY-1115"}
